requirements.txt
pillow
datetime
imagehash